carbon double->bigdecimal

+++ src/main/java/me/hakyuwon/ecostep/dto/CarbonStatsDto.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class CarbonStatsDto {
     private Long userId;
-    private double carbonReduction;
+    private BigDecimal carbonReduction;
 }
